+++ src-kernel/console.c
@@ -32,7 +32,7 @@ printint(int xx, int base, int sign)
   static char digits[] = "0123456789abcdef";
   char buf[16];
   int i;
-  uint x;
+  uint32_t x;
 
   if(sign && (sign = xx < 0))
     x = -xx;
@@ -57,7 +57,7 @@ void
 cprintf(char *fmt, ...)
 {
   int i, c, locking;
-  unsigned long *argp;
+  uintptr_t *argp;
   char *s;
 
   locking = cons.locking;
@@ -67,7 +67,7 @@ cprintf(char *fmt, ...)
   if (fmt == 0)
     panic("null fmt");
 
-  argp = (unsigned long*)(void*)(&fmt + 1);
+  argp = (uintptr_t *)(void *)(&fmt + 1);
   for(i = 0; (c = fmt[i] & 0xff) != 0; i++){
     if(c != '%'){
       ttypecho(c, consputc);
@@ -109,7 +109,7 @@ void
 panic(char *s)
 {
   int i;
-  uint pcs[10];
+  uintptr_t pcs[10];
 
   cli();
   cons.locking = 0;
